Preliminary refactoring of collision handling
This refactoring adds a new MoveableObject class which takes over the
responsibilities of moving an object independent of the controller.

Also, the moveable object knows which platform it's on.  This will allow
for taking the angle of the platform into account when moving on it.
Additionally, this will allow for platforms to impart special behavior
when they are collided with (e.g. adding velocity, doing damage, etc.).